Don't wait after SIGKILL when freeze-killing
When freeze-killing (for package force-stops) we attempt to wait until all
processes have 1) Received the kill signal AND 2) have actually exited and been
cleaned up AND 3) the associated cgroup(s) have been removed. The second and
third parts can take a long time, which can lead to lock contention in
ActivityManagerService when many kills are attempted simultaneously. For
freeze-kills with SIGKILL we do not need to wait to prevent app-restart
exploits. Now we send these processes a SIGKILL after freezing, but
do not wait until they are actually dead before returning.

Bug: 274646058
Test: Force-stop of chrome with 15 tabs completes ~500ms faster
Test: Full Play store update causes no ANR
Change-Id: Ib4cc2e31190338b565e4a89acce64d9ae52b19c9