Update FgsManager UI information on bg thread without lock
ActivityManagerService blocks on an awaiting incoming call while at the
same time requiring that the lock be held to provide information used to
make UI decisions. Here we copy data about which tasks are running and
who the current profiles are under then lock then post to the background
thread to load any needed UI data before updating the dialog on the main
thread.

Test: FgsManagerControllerTest
Bug: 276882028
Change-Id: I53664adbb534b05de8b130ed06fca0535f683156