Revert^2 "Show bouncer after user switch"
Caution: There is different behavior based on if you switch to or from the owner. Also there are various race conditions in the KeyguardViewMediator queue caused by waiting on async layouts to happen.

This CL tries to mitigate this issue by adding a delay, but more thorough investigation and cleanup of incoming signals and side effects is necessary to remove jank and make the queues order of execution predictable.

Test: Manually tested various user switch scenarios with different bouncers. PIN/SWIPE/NONE/FINGER.
Bug: b/266711330

366c6140a77cf93aae2ff25adb7a9ef27fde6d14

Change-Id: I4143f9727e870074a5a3bddf9e2586781f10b637