Fixes foldable autorotation setting being out of sync in QS and Settings
Both QS (SystemUI) and Settings, were using the same class to update
foldable auto-rotation settings. Since both are in different processes,
they have different internal memory state.

There was a bug where the internal state for the persisted setting was
only changed in the process that changed the setting.
When then trying to change the setting in the other process, it didn't
work as it thought that the setting was the same as before.

Test: DeviceStateRotationLockSettingsManagerTest.java
Test: Manually - Change settings in both QS and Settings and read logs.
Fixes: 279603743
Change-Id: I3f28027d5e24796ceb01fc1eb164aac8814a1753
Merged-In: I3f28027d5e24796ceb01fc1eb164aac8814a1753