Determine HEIF support by checking HEVC
ImageDecoder.isMimeTypeSupported always returns true for
HEIF and HEIC, however, according to CDD, this support is
optional, and they are only supported when the HEVC is
supported by the device.

This change is to fix inconsistency with CDD 5.1.5, and won't
impact handhelds or TV, as HEVC is a must for them. For auto
and watch, HEVC is not supported at all. So this change is safe.

Bug: 278664575
Test: atest BitmapFactoryTest ImageDecoderTest

Change-Id: Iecceb6f2a498fddc045ccce331f6fcef5debd2f5