Extends the CandidateAverageMetric with Auth Entry
This adds the auth entries response collective into the candidate
average metric, thereby creating a full aggregate for the response
information. It also sets up the Get emit logic, and is a step towards
final completion of the collection system.

The threaded timed get emit system was removed in favor of an immediate
function-based buffer/queue logging system that will be added once all
changes are in.

Bug: 271135048
Test: Build and Won't Submit without E2E Test

Change-Id: I17ac83de23dee1486a10d7ce02bfc334ceef5581