Add support for Input ads from the SDK runtime.
Symtom: EditText created from the SDK runtime is not able to accept
input using the soft keyboard. The root cause is a failed check in
ImeService when it tries to match the caller uid with the uid PM
associates with the sdk sandbox package. Since SDK sandbox package can
be running in a process that can have a UID from a range, rather than a
single UID, this check often fails.

Updated the check to use a PackageManager method that understands this
nuance about sdk sandbox package and ensures the UID falls within the
range of sdk sandbox UIDs.

Bug: b/242985591
Test: Using the SDK sandbox manual test app.
Change-Id: I3da937bf379f39c8204db5b7d7d1c41c6a6e942e