Don't register AdServices as dumpable anymore.
AdServicesManager registers the adservices_manager service, but it
it was temporary not doing so on Android 13 disabled as it was breaking
some devices that explicitly allowlisted services (on selinux). but starting on Android 14, the adservices_manager is guaranteed (by CTS)
to be published, so this workaround is not needed anymore.

Bug: 282239822

Test: adb shell dumpsys system_server_dumper --list > /dev/null |grep AdServices && echo FAIL || echo SUCCESS
Test: adb shell service check adservices_manager | grep "not found" && echo FAIL || echo SUCCESS

Change-Id: Ib81dd850c12c461cbb2515d56cec0d992d366642