Don't prematurely apply finishT in taskview handler
The finish-transaction is supposed to be applied by the
transition player since the order matters. By applying it
in TaskViewTransitions, it was causing operations to run
out-of-order leaving broken surface hierarchies. This
removes the erroneous apply() calls.

Also, set surface position to 0,0 on surfaceChanged to
match surfaceCreated -- otherwise we end up with a double
offset.

Bug: 277759491
Test: open new homecontrols from dream (observe no offset),
      then close them via the X button (should return to dream).
Change-Id: I4527df33b29ad93045928d8e73b0698a36d95081