Support non-managed profile Status Bar icons
When a managed profile's app is in the foreground, the status bar shows
a work briefcase icon to let the user know that the app is running in
the work profile.

However, this only works for managed (work) profiles, not for other
profile types.

As we explore other profile types, we need this same status bar
indicator. We therefore move the icon from SystemUi to Frameworks, tie
it to the profile type, and thereby allow non-managed profiles to also
display an icon.

For prod, this cl is a no-op, since only managed profiles are supported
in out-of-the-box AOSP. But using adb, other profile types can be
created, and they now can display a status bar icon.

Bug: 170249807
Bug: 274838657
Test: atest UserManagerServiceUserTypeTest
Test: atest com.android.server.pm.UserManagerTest#testProfileTypeInformation
Test: Manually confirmed that managed profiles have no change
Test: adb root; adb shell setprop persist.debug.user_mode_emulation 1 && adb shell pm create-user --user-type android.os.usertype.profile.TEST --profileOf 10 Testname  and verify that flask icon appears
Change-Id: Ic2099748b8cdb838730def886c46d990b068652b