Dismiss dialogs with the Escape key
While investigating what the Escape key is used for, I noticed how
it is used in dialogs in a very similar way as the Back key.

However, I don't think this captures the intent of the change, as
an unhandled back will likely be handled elsewhere in a way that
also dismisses the dialog.

This CL changes this and explicitly makes the Escape key dismiss
the dialog, cancelling it if possible.

Bug: 281809692
Test: Confirmed that the Private DNS dialog in settings is closed
properly.

Change-Id: I2334ee380e32751f0c1d93b9bbdf40d7d2d559fe