Fix issue with last auth entry
The problem is that we send an exception to the client and then send the pending data to the UI. As part of sending the exception, we also finish the session and in that finish logic, we are cancelling the pending intent. The status update is async so there's race condition. For now, I'm removing the cancellation of the pending intent so as part of the finish logic. It is still cancelled when a new pending intent is created for the same session but won't be cancelled when the session ends. May send out a follow up that addresses cancelling the pending intent with session finish in a safe way.

Bug: 281615828
Test: built and deployed locally

Change-Id: I9f30a0236da2125b56d8de42217fa70cc5e65191