Fix misplaced surface from fold -> unfold while bubbles are expanded
This also fixes the surface being misplaced when the theme changes.

When setBounds is called in TaskViewTransitions we were skipping
creating a new transition whenever there was anything pending.

Instead, we should only skip if the pending in flight is for that
task view AND an opening type transit, otherwise the bounds never
get updated until the next open / if a bounds change is sent without
any other pending.

Test: manual - have an expanded bubble on folded device, unfold
             => ensure the expanded view is fully visible
             - have an expanded bubble & change the theme via QS
             => ensure the expanded view is properly placed
             - test various bubble animations: expand, collapse,
               switch, new bubble, dismissing while expanded, rotation
               and ensure they all look fine.
Test: atest TaskViewTransitionTest
Bug: 281855397
Bug: 280440749
Change-Id: I43cbb90bdfbe90bf068c5de21d060193aa38c96f